+++ imap_tools/__init__.py
@@ -1,7 +1,7 @@
-from .query import Q, AND, OR, NOT, H
+from .query import AND, OR, NOT, Header, A, O, N, H
 from .mailbox import *
 from .message import *
 from .folder import *
 from .utils import *
 
-__version__ = '0.16.1'
+__version__ = '0.17.0'